bundler: cap inline-sourcemap source names; isolate chain tests
- InputSourceMap: reject maps whose sources[i] exceeds MAX_PATH_BYTES so
  an adversarial inline map can't panic the fixed-size path buffers;
  the map falls back cleanly instead.
- write_sources_for: resolve inner names via join_abs_string_buf_checked
  and emit the raw name when the join overflows, rather than the
  panicking unchecked join.
- Move the inline-sourcemap chain suite to its own test file and add an
  oversized-name regression test.
